secp256k1: Separate const time compare funcs.
The various functions for comparing in constant time are no longer
limited to only being used in mod n scalars, so move them into their own
file.
