+++ dcrec/secp256k1/modnscalar.go
@@ -210,47 +210,6 @@ func (s *ModNScalar) SetInt(ui uint32) *ModNScalar {
 	return s
 }
 
-// constantTimeEq returns 1 if a == b or 0 otherwise in constant time.
-func constantTimeEq(a, b uint32) uint32 {
-	return uint32((uint64(a^b) - 1) >> 63)
-}
-
-// constantTimeEq64 returns 1 if a == b or 0 otherwise in constant time.
-func constantTimeEq64(a, b uint64) uint32 {
-	t := a ^ b
-	return uint32(((t | -t) >> 63) ^ 1)
-}
-
-// constantTimeNotEq returns 1 if a != b or 0 otherwise in constant time.
-func constantTimeNotEq(a, b uint32) uint32 {
-	return ^uint32((uint64(a^b)-1)>>63) & 1
-}
-
-// constantTimeLess returns 1 if a < b or 0 otherwise in constant time.
-func constantTimeLess(a, b uint32) uint32 {
-	return uint32((uint64(a) - uint64(b)) >> 63)
-}
-
-// constantTimeLessOrEq returns 1 if a <= b or 0 otherwise in constant time.
-func constantTimeLessOrEq(a, b uint32) uint32 {
-	return uint32((uint64(a) - uint64(b) - 1) >> 63)
-}
-
-// constantTimeGreater returns 1 if a > b or 0 otherwise in constant time.
-func constantTimeGreater(a, b uint32) uint32 {
-	return constantTimeLess(b, a)
-}
-
-// constantTimeGreaterOrEq returns 1 if a >= b or 0 otherwise in constant time.
-func constantTimeGreaterOrEq(a, b uint32) uint32 {
-	return constantTimeLessOrEq(b, a)
-}
-
-// constantTimeMin returns min(a,b) in constant time.
-func constantTimeMin(a, b uint32) uint32 {
-	return b ^ ((a ^ b) & -constantTimeLess(a, b))
-}
-
 // overflows determines if the current scalar is greater than or equal to the
 // group order in constant time and returns 1 if it is or 0 otherwise.
 func (s *ModNScalar) overflows() uint32 {
